refactor(testkit-js): replace the @/ path alias with relative imports
The `@/*` -> `./src/*` alias made internal modules look like bare package
specifiers. rollup-plugin-dts does not honour tsconfig `paths`, so it left them
untouched: the published `dist/index.d.ts` carried
`import ... from '@/test-environment'`, which no consumer can resolve. Those
declarations were already broken.

The alias also blocks the only correct rule for a library bundle - a specifier
that is not relative or absolute is a runtime dependency - because `@/utils`
looks bare while being internal.

Mechanical rewrite of 45 specifiers across 35 files, plus removal of the alias
from both tsconfigs. `vite-tsconfig-paths` existed only to resolve it, so the
plugin and its devDependency go too.
