+++ testkit-js/testkit-js/package.json
@@ -65,7 +65,6 @@
     "rxjs": "^7.8.2",
     "typedoc": "^0.28.19",
     "typedoc-plugin-markdown": "4.12.0",
-    "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^4.1.7"
   },
   "files": [
