+++ testkit-js/testkit-js/src/client/indexer-client.ts
@@ -16,7 +16,7 @@
 import axios from 'axios';
 import type { Logger } from 'pino';
 
-import { buildUrlWithPath } from '@/utils';
+import { buildUrlWithPath } from '../utils';
 
 export class IndexerClient {
   readonly indexerUrl: string;
