+++ testkit-js/testkit-js/src/client/node-client.ts
@@ -20,7 +20,7 @@ import type { BlockHash } from '@midnight-ntwrk/midnight-js-types';
 import axios, { type AxiosResponse } from 'axios';
 import type { Logger } from 'pino';
 
-import { buildUrlWithPath } from '@/utils';
+import { buildUrlWithPath } from '../utils';
 
 /**
  * Client for interacting with a Midnight node's JSON-RPC API
