+++ testkit-js/testkit-js/src/test-environment/test-environments/preview-test-environment.ts
@@ -15,8 +15,7 @@
 
 import { NetworkId } from '@midnightntwrk/wallet-sdk';
 
-import type { EnvironmentConfiguration } from '@/test-environment';
-
+import type { EnvironmentConfiguration } from '..';
 import { RemoteTestEnvironment } from './remote-test-environment';
 
 /**
