+++ testkit-js/testkit-js/src/test-environment/test-environments/qanet-test-environment.ts
@@ -14,8 +14,7 @@
  */
 import { NetworkId } from '@midnightntwrk/wallet-sdk';
 
-import type { EnvironmentConfiguration } from '@/test-environment';
-
+import type { EnvironmentConfiguration } from '..';
 import { RemoteTestEnvironment } from './remote-test-environment';
 
 /**
