+++ testkit-js/testkit-js/src/test-environment/test-environments/test-environment.ts
@@ -15,9 +15,8 @@
 
 import type { Logger } from 'pino';
 
-import type { ProofServerContainer } from '@/proof-server-container';
-import type { MidnightWalletProvider } from '@/wallet';
-
+import type { ProofServerContainer } from '../../proof-server-container';
+import type { MidnightWalletProvider } from '../../wallet';
 import type { EnvironmentConfiguration } from '../environment-configuration';
 
 /**
