+++ testkit-js/testkit-js/vitest.config.ts
@@ -13,11 +13,9 @@
  * limitations under the License.
  */
 
-import tsconfigPaths from 'vite-tsconfig-paths';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  plugins: [tsconfigPaths()],
   test: {
     globals: true,
     environment: 'node',
